+++ NEWS.md
@@ -5,6 +5,7 @@
 
 #### Highlights
 * Fix ReadObs() so date/time format works when importing as a data.table object
+* Prevent scientific notation when writing files
 
 ## HYPEtools 1.5.1 (2023-08-24)
 *Enhancements and bug fixes*